# openapi/glass-game-panel-v1.openapi.yaml
openapi: 3.1.0
info:
  title: Glass Game Panel Control Plane API
  version: 0.1.5-stub
  summary: GlassGamePanel control plane (draft)
  description: |
    Draft control-plane API for GlassGamePanel (`0.1.5-stub`). Humans and agents
    share this API. A UI is another client. Wings is the current executor; a
    later driver may sit behind the same verbs. This draft is not a production deploy.

    **Defaults:**
    - Documented production base: `https://panel-api.glasshosting.com/v1` (not published with this draft)
    - `serverId` is a Glass UUID. The Wings UUID is `wings_server_id`
    - Auth is Portal OIDC. Portal mints agent tokens. Agent tokens cannot mint child tokens
    - `agents:manage` is for people, not for agent tokens
    - Console is a ticket plus WebSocket. No Authorization header on the WebSocket handshake
    - Agents use the files API. Optional SFTP for people is later
    - RCON is deferred to v1.1
    - `network` is reserved. There is no allocate or release path in this draft
    - Rate limits: 60 mutate requests and 300 read requests per minute per token
    - Repository: `FuzzySpace/GlassGamePanel`

    **Migration:**
    - GlassHosting-owned Pterodactyl first. A panel the customer operates is later
    - Default mode is `inventory_only`
    - Rollback window default is 72 hours (design stub)
    - Servers the caller does not own stay excluded. `include_managed_inventory` is forced false in this build

    **Refuse:**
...
    not handed to agents.

...

...

...
servers:
...
paths:
  /servers:
    get:
      tags: [Servers]
      operationId: listServers
      summary: List servers visible to the caller
      security:
        - BearerAuth: [servers:read]
      parameters:
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Limit'
        - name: status
          in: query
          schema:
            $ref: '#/components/schemas/ServerStatus'
      responses:
        '200':
          description: Cursor page of servers
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServerPage'
        default:
          $ref: '#/components/responses/Error'
    post:
      tags: [Servers]
      operationId: createServer
      summary: Create a server (job completes with an IPv4 and IPv6 claim; refused servers return 403 managed_inventory_denied; default executor does not dial Wings)
      security:
        - BearerAuth: [servers:write]
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ServerCreate'
      responses:
        '202':
          description: Accepted; provision job started
          headers:
            Location:
              description: Job URL
              schema:
                type: string
                format: uri
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Job'
        default:
          $ref: '#/components/responses/Error'
...
  /migrations:
    get:
      tags: [Migrations]
      operationId: listMigrations
      summary: List migration jobs visible to the caller
      security:
        - BearerAuth: [migrations:read]
      parameters:
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Limit'
        - name: status
          in: query
          schema:
            $ref: '#/components/schemas/MigrationStatus'
      responses:
        '200':
          description: Migration page
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MigrationPage'
        default:
          $ref: '#/components/responses/Error'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Short-lived JWT from GlassPortal OIDC. Scopes are space-delimited.
        Portal mints agent tokens (`POST /agent-tokens` is a staff or owner
        session). Agents cannot mint child tokens. Do not use long-lived
        administrator keys. Mutual TLS between Portal and Wings is transport
        and is outside this OpenAPI file.
  parameters:
...
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema:
        type: string
        minLength: 8
        maxLength: 128
      description: |
        Required on all mutating requests. Same key + same body within 24h
        returns the original result (safe agent retry).
    Cursor:
      name: cursor
      in: query
      schema:
        type: string
      description: Opaque cursor from prior page
    Limit:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 25
  responses:
    Error:
      description: Stable error envelope
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message, request_id]
          properties:
            code:
              type: string
              description: Stable machine code
              enum:
                - unauthorized
                - forbidden
                - managed_inventory_denied
                - migration_source_unreachable
                - migration_managed_inventory_excluded
                - migration_remap_conflict
                - migration_rollback_expired
                - not_found
                - conflict
                - validation_failed
                - rate_limited
                - idempotency_replay
                - executor_unavailable
                - internal
            message:
              type: string
            request_id:
              type: string
            details:
              type: object
              additionalProperties: true
    ServerStatus:
      type: string
      enum: [pending, installing, offline, starting, running, stopping, stopping_force, suspended, deleted]
    Server:
      type: object
      required: [id, name, status, egg_id, node_id, created_at]
      properties:
        id:
          type: string
          format: uuid
          description: Glass server UUID (SoT primary id)
        wings_server_id:
          type: [string, 'null']
          format: uuid
          description: Wings/Ptero server UUID ref (phase A executor); null until provisioned
        name:
          type: string
        status:
          $ref: '#/components/schemas/ServerStatus'
        egg_id:
          type: string
        node_id:
          type: string
          description: Wings node id (phase A)
        entitlement_id:
          type: [string, 'null']
        alloc_pool:
          type: [string, 'null']
          description: Named allocation pool. The configured pool is the only one this build accepts.
        allocations:
          type: array
          items:
            $ref: '#/components/schemas/Allocation'
        limits:
          $ref: '#/components/schemas/ResourceLimits'
        tags:
          type: array
          items:
            type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    ServerPage:
      type: object
      required: [data]
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Server'
        next_cursor:
          type: [string, 'null']
    ServerCreate:
      type: object
      required: [name, egg_id, entitlement_id]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 64
        egg_id:
          type: string
        entitlement_id:
          type: string
          description: Must be active Portal entitlement owned by caller
        node_id:
          type: string
          description: Optional pin; else scheduler
        limits:
          $ref: '#/components/schemas/ResourceLimits'
        environment:
          type: object
          additionalProperties:
            type: string
          description: Egg env vars (allowlisted keys only)
...
    Allocation:
      type: object
      required: [ip, port]
      properties:
        ip:
          type: string
          description: |
            IPv4 or IPv6 literal. IPv6 rows from the configured pool are
            local addresses, not a public AAAA.
        port:
          type: integer
          minimum: 1
          maximum: 65535
        family:
          type: string
          enum: [ipv4, ipv6]
          description: Address family. A Phase A create claims one of each.
        routable:
          type: boolean
          description: False when the IPv6 row is a local address and not a public AAAA.
        notes:
          type: string
        ip_alias:
          type: string
        node_id:
          type: string
        alloc_pool:
          type: string
    ResourceLimits:
      type: object
      properties:
        memory_mb:
          type: integer
          minimum: 128
        disk_mb:
          type: integer
          minimum: 1024
        cpu_percent:
          type: integer
          minimum: 10
          maximum: 800
        databases:
          type: integer
          minimum: 0
        backups:
          type: integer
          minimum: 0
...
    MigrationStatus:
      type: string
      enum:
        - queued
        - discovering
        - planned
        - awaiting_confirm
        - copying
        - cutting_over
        - succeeded
        - partial
        - failed
        - rolling_back
        - rolled_back
        - cancelled
...
    MigrationPage:
      type: object
      required: [data]
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Migration'
        next_cursor:
          type: [string, 'null']
...
    Job:
      type: object
      required: [id, type, status, created_at]
      properties:
        id:
          type: string
          format: uuid
        type:
          type: string
          enum: [server.create, server.delete, power, backup.create, backup.restore, install, migration.ptero, migration.rollback]
        status:
          type: string
          enum: [queued, running, succeeded, failed, cancelled]
        server_id:
          type: [string, 'null']
          format: uuid
        progress_percent:
          type: [integer, 'null']
          minimum: 0
          maximum: 100
        error_code:
          type: [string, 'null']
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        completed_at:
          type: [string, 'null']
          format: date-time
        wings_server_id:
          type: [string, 'null']
          description: Null while the Wings executor is noop.
        allocations:
          type: array
          description: Present on a succeeded server.create. One IPv4 row and one IPv6 row.
          items:
            $ref: '#/components/schemas/Allocation'
        result:
          type: object
          description: |
            Present when server.create succeeds. `ula_routable` is false and
            `not_public_aaaa` is true for the configured pool.
          properties:
            id:
              type: string
              format: uuid
            glass_server_id:
              type: string
              format: uuid
            wings_server_id:
              type: [string, 'null']
            node_id:
              type: string
            alloc_pool:
              type: string
            ula_prefix:
              type: string
            ula_routable:
              type: boolean
            not_public_aaaa:
              type: boolean
            executor:
              type: string
            wings_dispatch:
              type: boolean
            entitlement_id:
              type: string
            allocations:
              type: array
              items:
                $ref: '#/components/schemas/Allocation'

# Scope catalog for agents and people
...